+++ src/main/java/io/github/aquerr/eaglefactions/storage/IPlayerStorage.java
@@ -10,19 +10,19 @@ public interface IPlayerStorage
 {
     boolean checkIfPlayerExists(UUID playerUUID, String playerName);
 
-    boolean addPlayer(UUID playerUUID, String playerName, BigDecimal startingPower, BigDecimal maxPower);
+    boolean addPlayer(UUID playerUUID, String playerName, float startingPower, float maxPower);
 
     boolean setDeathInWarzone(UUID playerUUID, boolean didDieInWarZone);
 
     boolean getLastDeathInWarzone(UUID playerUUID);
 
-    BigDecimal getPlayerPower(UUID playerUUID);
+    float getPlayerPower(UUID playerUUID);
 
-    boolean setPlayerPower(UUID playerUUID, BigDecimal power);
+    boolean setPlayerPower(UUID playerUUID, float power);
 
-    BigDecimal getPlayerMaxPower(UUID playerUUID);
+    float getPlayerMaxPower(UUID playerUUID);
 
-    boolean setPlayerMaxPower(UUID playerUUID, BigDecimal maxpower);
+    boolean setPlayerMaxPower(UUID playerUUID, float maxpower);
 
     Set<String> getServerPlayerNames();
